requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
aiohttp==3.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
